Resolve exception complaining about OptiX context validation failure
— OptiX context validation is no longer performed immediately after
initialization of the context as this was causing exception complaining
that there was no ray generation shaders attached to the context at that
point (which was expected as they were getting attached later). Instead,
the validation is now performed at the very final stage, right before
launching the ray tracing

— Project properties altered to automatically copy OptiX DLL into the
target compilation directory

— Slightly modified demo/gomos bat launch script hoping it would finally
work outside of visual studio. It gets there, but still silently fails
due to some uncaught (or eaten?) exception